refactor: extract hasConversationWith to twilioUtils
squash: log

squash: unused import
